Rename config and logging files
